CVE checker: Update remarks for latest version
The mnemonic for the remark Ignored has been split into False Positive
and Not Affected in the recent version of cve_bin_tool, preventing it
from properly parsing the CVS file and not being able to do proper
report.

Use Not Affected, as it is seems to be the closest to what Ignored was
used for in our case.

Pull Request: https://projects.blender.org/blender/blender/pulls/125645